+++ src/host/pulseaudio/stream.rs
@@ -10,9 +10,9 @@ use futures::executor::block_on;
 use pulseaudio::{protocol, AsPlaybackSource};
 
 use crate::{
-    traits::StreamTrait, BackendSpecificError, BuildStreamError, Data, InputCallbackInfo,
-    InputStreamTimestamp, OutputCallbackInfo, OutputStreamTimestamp, PlayStreamError, SampleFormat,
-    StreamError, StreamInstant,
+    traits::StreamTrait, BackendSpecificError, BuildStreamError, Data, FrameCount,
+    InputCallbackInfo, InputStreamTimestamp, OutputCallbackInfo, OutputStreamTimestamp,
+    PlayStreamError, SampleFormat, StreamError, StreamInstant,
 };
 
 pub enum Stream {
@@ -43,6 +43,22 @@ impl StreamTrait for Stream {
 
         res.map_err(Into::<BackendSpecificError>::into)?;
         Ok(())
+    }
+
+    fn buffer_size(&self) -> Option<FrameCount> {
+        let (spec, bytes) = match self {
+            Stream::Playback(s) => (
+                s.sample_spec(),
+                s.buffer_attr().minimum_request_length as usize,
+            ),
+            Stream::Record(s) => (s.sample_spec(), s.buffer_attr().fragment_size as usize),
+        };
+        let frame_size = spec.channels as usize * spec.format.bytes_per_sample();
+        if bytes > 0 {
+            Some((bytes / frame_size) as _)
+        } else {
+            None
+        }
     }
 }